chatbot: separate main.js in three main functions
Create function  `buildConnector`, `buildBot` and `buildBotApi`.
`buildConnector` is responsible for create a chat connector. `buildBot`
is responsible for create a bot, that is, the function responsible for
receive and reply messages and `buildBotApi` is responsible for create a
restful API using restify and configure `/api/messages` route to
listening using bot.
