+++ app/main.js
@@ -1,25 +1,13 @@
 import env from 'node-env-file'
-import restify from 'restify'
-import {
-  ChatConnector,
-  UniversalBot,
-} from 'botbuilder'
+import Promise from 'bluebird'
 import { basePath } from '../config/paths'
+import buildConnector from './buildConnector'
+import buildBot from './buildBot'
+import buildBotApi from './buildBotApi'
 
-env(basePath() + '/.env')
+env(`${basePath()}/.env`)
 
-const server = restify.createServer()
-server.listen(process.env.PORT || 3987, () => {
-  console.log('% listening on %s', server.name, server.url)
-})
-
-const connector = new ChatConnector({
-  appId: process.env.MICROSOFT_APP_ID,
-  appPassword: process.env.MICROSOFT_APP_PASSWORD,
-})
-
-server.post('/api/messages', connector.listen())
-
-const bot = new UniversalBot(connector, session => {
-  session.send('You said %s', session.message.text)
-})
+Promise.resolve()
+  .then(buildConnector)
+  .then(buildBot)
+  .then(buildBotApi)
